+++ fceulib/trace.cc
@@ -69,7 +69,7 @@ vector<uint8> RLE::CompressEx(const vector<uint8> &in,
   // 255.
   const int max_antirun_length = (int)(255 - run_cutoff) + 1;
 
-  for (int i = 0; i < in.size(); /* in loop */) {
+  for (int i = 0; i < (int)in.size(); /* in loop */) {
     // Greedy: Grab the longest prefix of bytes that are the same,
     // up to max_run_length.
     const uint8 target = in[i];
@@ -77,7 +77,7 @@ vector<uint8> RLE::CompressEx(const vector<uint8> &in,
     // that this is legal.
     int run_length = 1;
     while (run_length < max_run_length &&
-           i + run_length < in.size() &&
+           i + run_length < (int)in.size() &&
            in[i + run_length] == target) {
       run_length++;
     }
@@ -98,7 +98,7 @@ vector<uint8> RLE::CompressEx(const vector<uint8> &in,
       // same.
       int anti_run_length = 1;
       while (anti_run_length < max_antirun_length &&
-             i + anti_run_length + 1 < in.size() &&
+             i + anti_run_length + 1 < (int)in.size() &&
              in[i + anti_run_length] !=
              in[i + anti_run_length + 1]) {
         anti_run_length++;
@@ -128,14 +128,14 @@ bool RLE::DecompressEx(const vector<uint8> &in,
                        vector<uint8> *out) {
   out->clear();
 
-  for (int i = 0; i < in.size(); /* in loop */) {
+  for (int i = 0; i < (int)in.size(); /* in loop */) {
     const uint8 control = in[i];
     i++;
     if (control <= run_cutoff) {
       // If less than the run cutoff, we treat it as a run.
       const int run_length = control + 1;
 
-      if (i >= in.size()) {
+      if (i >= (int)in.size()) {
         return false;
       }
 
@@ -153,7 +153,7 @@ bool RLE::DecompressEx(const vector<uint8> &in,
       // represented as 0) so we code starting at 2.
       const int antirun_length = control - run_cutoff + 1;
 
-      if (i + antirun_length >= in.size()) {
+      if (i + antirun_length >= (int)in.size()) {
         return false;
       }
 
@@ -274,8 +274,8 @@ string Traces::Difference(const Trace &l, const Trace &r) {
     int num_differences = 0;
     int first_difference = -1;
     // Long version.
-    for (int i = 0; i < min(l.data_memory.size(),
-                            r.data_memory.size()); i++) {
+    for (int i = 0; i < (int)min(l.data_memory.size(),
+                                 r.data_memory.size()); i++) {
       if (!(i % 16)) how += "\n";
       uint8 ll = l.data_memory[i], rr = r.data_memory[i];
       if (ll != rr) {
@@ -293,8 +293,8 @@ string Traces::Difference(const Trace &l, const Trace &r) {
     return how;
 
     // Short version.
-    for (int i = 0; i < min(l.data_memory.size(),
-                            r.data_memory.size()); i++) {
+    for (int i = 0; i < (int)min(l.data_memory.size(),
+                                 r.data_memory.size()); i++) {
       uint8 ll = l.data_memory[i], rr = r.data_memory[i];
       if (ll != rr) {
         if (how.length() > 70)
@@ -442,7 +442,7 @@ vector<Traces::Trace> Traces::ReadFromFile(const string &filename) {
       }
       // fprintf(stderr, "String of length %u.\n", len);
       t.data_string.resize(len);
-      for (int i = 0; i < len; i++) {
+      for (uint32 i = 0; i < len; i++) {
         if (!Read8((uint8*)&t.data_string[i])) {
           fprintf(stderr, "Incomplete string.\n");
           abort();
@@ -458,7 +458,7 @@ vector<Traces::Trace> Traces::ReadFromFile(const string &filename) {
       }
       vector<uint8> compressed;
       compressed.resize(len);
-      for (int i = 0; i < len; i++) {
+      for (uint32 i = 0; i < len; i++) {
         if (!Read8(&compressed[i])) {
           fprintf(stderr, "Incomplete memory.\n");
           abort();
